safari-534.53-branch/LayoutTests/fast/dom/collection-nameditem-move-between-documents-expected.txt

PASS typeof children.namedItem('child') is 'undefined'
PASS successfullyParsed is true

TEST COMPLETE
